+++ src/constants.ts
@@ -1,0 +1,9 @@
+/* ========================================================================== */
+/*                                  CONSTANTS                                 */
+/* ========================================================================== */
+
+export const EXT_NAME = 'Comment Divider';
+export const EXT_ID = 'comment-divider';
+
+export const GAP_SYM = ' ';
+export const NEW_LINE_SYM = '\n';