+++ src/limiters.ts
@@ -1,0 +1,102 @@
+/* ========================================================================== */
+/*                             LIMITERS COLLECTION                            */
+/* ========================================================================== */
+
+export interface ILimiters {
+  left: string;
+  right: string;
+}
+
+///
+
+const wrapLimiters = (left: string, right: string): ILimiters => ({ left, right });
+
+///
+
+export const getLanguageLimiters = (lang?: string): ILimiters => {
+  switch (lang) {
+    case 'c':
+    case 'cpp':
+    case 'csharp':
+    case 'css':
+    case 'go':
+    case 'groovy':
+    case 'java':
+    case 'javascript':
+    case 'javascriptreact':
+    case 'jsonc':
+    case 'kotlin':
+    case 'less':
+    case 'objective-c':
+    case 'php':
+    case 'sass':
+    case 'scala':
+    case 'stylus':
+    case 'sql':
+    case 'swift':
+    case 'typescript':
+    case 'typescriptreact':
+      return wrapLimiters('/*', '*/');
+
+    case 'bash':
+    case 'dockerfile':
+    case 'coffeescript':
+    case 'ignore':
+    case 'julia':
+    case 'makefile':
+    case 'perl':
+    case 'perl6':
+    case 'powershell':
+    case 'properties':
+    case 'python':
+    case 'r':
+    case 'ruby':
+    case 'shell':
+    case 'shellscript':
+    case 'yaml':
+    case 'yml':
+      return wrapLimiters('#', '#');
+
+    case 'html':
+    case 'markdown':
+    case 'plist':
+    case 'xaml':
+    case 'xml':
+    case 'xsl':
+      return wrapLimiters('<!--', '-->');
+
+    case 'clojure':
+    case 'lisp':
+    case 'scheme':
+      return wrapLimiters(';', ';');
+
+    case 'elm':
+    case 'haskell':
+    case 'lua':
+      return wrapLimiters('--', '--');
+
+    case 'erlang':
+    case 'latex':
+    case 'matlab':
+      return wrapLimiters('%', '%');
+
+    case 'jade':
+    case 'pug':
+      return wrapLimiters('//-', '-//');
+
+    case 'fsharp':
+      return wrapLimiters('(*', '*)');
+
+    case 'bat':
+      return wrapLimiters('REM', '');
+
+    case 'vb':
+      return wrapLimiters("'", "'");
+
+    case 'plaintext':
+      return wrapLimiters('#', '#');
+
+    default:
+      return wrapLimiters('/*', '*/');
+  }
+};